Provide updateCellMeasurements() method for cell height refresh
While the table height is being updated at each update that
gives a different sum of rows heights, sometimes the individual
cell height is not being updated.

Since this depends on some obscure logic for the CellMeasurer cache
that maintains the cell previously measurements, one way of
refreshing those measurements is to provide the `measure()` render
prop provided by the `CellMeasurer`. That way one can manually
remeasure an individual cell and update the entire row.